creating custom Logger with support for .notice()
