+++ python/bin/open_spimescape_portal
@@ -10,7 +10,7 @@ import yaml
 import dripline
 
 import logging
-logger = logging.getLogger(__name__)
+logger = dripline.core.DriplineLogger(__name__)#logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
 
@@ -26,7 +26,7 @@ def open_spimescape_portal(nodename, broker, **kwargs):
     logger.info('spimescapes created and populated')
     #logger.info('creating bindings')
     #portal.create_bindings()
-    logger.info('starting consumption')
+    logger.notice('Configuration of {} complete, starting consumption'.format(nodename))
     portal.start_event_loop()
 
 def create_child(portal, conf_dict):
