+++ python/dripline/core/__init__.py
@@ -17,3 +17,4 @@ from .portal import *
 from .provider import *
 from .service import *
 from .spime import *
+from .status_logger import *
